+++ common/src/main/java/com/kingpixel/cobbleutils/Model/zones/zoneshapes/CuboidShape.java
@@ -1,11 +1,14 @@
 package com.kingpixel.cobbleutils.Model.zones.zoneshapes;
 
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.math.BlockPos;
 import org.jetbrains.annotations.Nullable;
 
+@EqualsAndHashCode(callSuper = true) @Data
 public class CuboidShape extends ZoneShape {
 
   public static final String TYPE = "CUBOID";